+++ src/parse.test.js
@@ -6,15 +6,22 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 import fs from 'fs';
-import { parseCptText } from './parse';
+import { parseCpt } from './parse';
 
 const PATH = __dirname + '/../fixtures';
 const FILENAMES = fs.readdirSync(PATH).filter(x => x.endsWith('.cpt'));
 
-describe('parseCptText', () => {
-  it.each(FILENAMES)('returns a color palette', (filename) => {
-    const text = fs.readFileSync(PATH + '/' + filename).toString();
-    const actual = parseCptText(text);
+describe('parseCpt', () => {
+  it.each(FILENAMES)('returns a color palette from a text', (filename) => {
+    const cptText = fs.readFileSync(PATH + '/' + filename).toString();
+    const actual = parseCpt(cptText);
+
+    expect(actual).toBeDefined();
+  });
+
+  it.each(FILENAMES)('returns a color palette from an array', (filename) => {
+    const cptArray = fs.readFileSync(PATH + '/' + filename.replace('.cpt', '.json')).toString();
+    const actual = parseCpt(cptArray);
 
     expect(actual).toBeDefined();
   });